refactor(redux): move calBondDetails-related funcs to utils
